refactor: use bookmark status hook functionality
The hook useBookmarkStatus was refactor and its functions, the following
changes have been made:
  - Updated sequence of operations in the hook, now isPaintingBookmarked
used only to determine the initial state of isBookmarked
  - toggleBookmark accepts payload as props type of number (for
deleting the card from the store) and PaintingCardInfoType (to add the
card to the store). Currently toggleBookmark dispatches deletePainting
and savePainting depending on the prev state
  - Updated setIsBookmarked to use prevState to ensure that state
updates correctly and avoids potential issues with stale state.
Additionally updated components that are using useBookmarkStatus hook
due to new usage of toggleBookmark function